+++ src/Contact.Domain/Contacts/Repositories/IContactRepository.cs
@@ -1,0 +1,8 @@
+﻿using Contacts.Domain.Contacts.Models;
+using Contacts.Infrastructure.Repositories;
+
+namespace Contacts.Domain.Contacts.Repositories;
+public interface IContactRepository : IRepositoryBase<Contact>
+{
+    Task<IEnumerable<Contact>> GetByDDDAsync(string ddd);
+}
