--- src/Contact.Domain/Contracts/Repositories/IContactRepository.cs
@@ -1,7 +1,0 @@
-﻿using Contacts.Domain.Entities;
-
-namespace Contacts.Domain.Contracts.Repositories;
-public interface IContactRepository : IRepositoryBase<Contact>
-{
-    Task<IEnumerable<Contact>> GetByDDDAsync(string ddd);
-}
